django/requirements.txt
appdirs==1.4.2
Django==1.10.5
djangorestframework==3.4.6
drfdocs==0.0.11
packaging==16.8
pyparsing==2.1.10
six==1.10.0
